decloupled the reading of the productions
The productions and lex tokens can be retrieved and the there is no
tight coupling between gocc and the application, in that regard.
the logic, however, still is. on that note, the old code still exists,
the new one is slow being added, but the old one is not being deleted
yet.